requirements.txt

_libgcc_mutex=0.1
apscheduler=3.6.3
blas=1.0
ca-certificates=2021.4.13
certifi=2020.12.5
cffi=1.14.5
cryptography=3.4.7
intel-openmp=2021.2.0
ld_impl_linux-64=2.33.1
libffi=3.3
libgcc-ng=9.1.0
libstdcxx-ng=9.1.0
mkl=2021.2.0
mkl-service=2.3.0
mkl_fft=1.3.0
mkl_random=1.2.1
ncurses=6.2
numpy=1.20.1
numpy-base=1.20.1
openssl=1.1.1k
pip=21.0.1
pycparser=2.20
pysocks=1.7.1
python=3.8.8
python-telegram-bot=13.4.1
pytz=2021.1
readline=8.1
setuptools=52.0.0
six=1.15.0
sqlite=3.35.4
tk=8.6.10
tornado=6.1
tzlocal=2.1
ujson=4.0.2
wheel=0.36.2
xz=5.2.5
zlib=1.2.11